Completed unit 4.7 - Demo: Casting

+++ c-plus-plus-fundamentals/module_03/module_03/Hello.cpp
@@ -20,7 +20,7 @@ int main()
 
 	double d1 = 2.2;
 	double d2 = i1;
-	int i5 = d1;
+	int i5 = static_cast<int>(d1);
 	cout << "d1 = " << d1 << endl;
 	cout << "d2 = " << d2 << endl;
 	cout << "i5 = " << i5 << endl;
@@ -47,8 +47,8 @@ int main()
 	auto a8 = 0xFF;
 	auto a9 = 0b111;
 
-	a1 = a2;
-
+	a1 = static_cast<int>(a2);
+	/*
 	unsigned char n1 = 128;
 	char n2 = 128;
 	cout << "n1 = " << n1 << endl;
@@ -61,6 +61,6 @@ int main()
 	n2 = 300;
 	cout << "n1 = " << n1 << endl;
 	cout << "n2 = " << n2 << endl;
-
+	*/
 	return 0;
 }
